[SPOT-281][REFACTOR] 회원 선호 지역, 분야 및 스터디 이유 관련 기능 유즈 케이스로 분리
